GSYE-295: Re-arrange market_utils.py into market_counters
