+++ src/gsy_e/gsy_e_core/global_objects_singleton.py
@@ -17,7 +17,7 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
 from gsy_e.gsy_e_core.user_profile_handler import ProfilesHandler
 from gsy_e.gsy_e_core.global_stats import ExternalConnectionGlobalStatistics
-from gsy_e.gsy_e_core.market_utils import FutureMarketCounter
+from gsy_e.gsy_e_core.market_counters import FutureMarketCounter
 
 
 class GlobalObjects:
