+++ src/gsy_e/models/area/__init__.py
@@ -32,8 +32,7 @@ from gsy_e.gsy_e_core.blockchain_interface import blockchain_interface_factory
 from gsy_e.gsy_e_core.device_registry import DeviceRegistry
 from gsy_e.gsy_e_core.exceptions import AreaException, GSyException
 from gsy_e.gsy_e_core.myco_singleton import bid_offer_matcher
-from gsy_e.gsy_e_core.util import TaggedLogWrapper
-from gsy_e.gsy_e_core.market_utils import is_external_matching_enabled
+from gsy_e.gsy_e_core.util import TaggedLogWrapper, is_external_matching_enabled
 from gsy_e.models.area.event_dispatcher import DispatcherFactory
 from gsy_e.models.area.events import Events
 from gsy_e.models.area.markets import AreaMarkets
